pyproject.toml
[project]
name = "colortheme"
version = "0.1.0"
description = "Change and backup colortheme configuration"
authors = [{name = "Quinten Roets", email = "qdr2104@columbia.edu"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "astral >= 3.2",
    "backupmaster >= 1.0.0",
    "geocoder >= 1.38.1",
]

[project.optional-dependencies]
dev = [
    "package-dev-tools >=0.5.11, <1",
    "package-dev-utils >=0.1.6, <1",
]

[project.urls]
"Source Code" = "https://github.com/quintenroets/colortheme"

[project.scripts]
colortheme = "colortheme.cli:entry_point"
restartplasma = "colortheme.cli:restart_plasma"
light = "colortheme.cli:go_light"
dark = "colortheme.cli:go_dark"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.coverage.run]
command_line = "-m pytest"

[tool.coverage.report]
precision = 4
fail_under = 100

[tool.mypy]
strict = true
no_implicit_reexport = false

[[tool.mypy.overrides]]
module = [
    "geocoder.*",
    "gui.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
pythonpath = [
    "src", ".",
]

[tool.ruff]
fix = true

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "UP",  # pyupgrade
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.setuptools.package-data]
colortheme = ["py.typed", "assets/**"]
